viewer を分割
