+++ nodejs/requestHandlers/epgrec.js
@@ -1,8 +1,8 @@
-var viewer = require(__dirname + "/../viewer");
+var viewerEpgrec = require(__dirname + "/../viewer/epgrec");
 var log = require(__dirname + "/../logger").getLogger();
 
 module.exports = function(response, parsedUrl) {
     log.access.info("Request handler 'epgrec' was called.");
-    viewer.epgrec(response);
+    viewerEpgrec(response);
 }
