+++ nodejs/requestHandlers/notFound.js
@@ -1,8 +1,8 @@
-var viewer = require(__dirname + "/../viewer");
+var viewerNotFound = require(__dirname + "/../viewer/notFound");
 var log = require(__dirname + "/../logger").getLogger();
 
 module.exports = function(response) {
     log.access.info("No request handler found.");
-    viewer.notFound(response);
+    viewerNotFound(response);
 }
