+++ nodejs/requestHandlers/tvProgram.js
@@ -1,4 +1,4 @@
-var viewer = require(__dirname + "/../viewer");
+var viewerTvProgram = require(__dirname + "/../viewer/tvProgram");
 var sqlModel = require(__dirname + "/../sqlModel");
 var log = require(__dirname + "/../logger").getLogger();
 
@@ -6,7 +6,7 @@ module.exports = function(response, parsedUrl) {
     log.access.info("Request handler 'tvProgram' was called.");
     sqlModel.getNowEpgData(function(results) {
         log.access.debug("called callback");
-        viewer.tvProgram(response, results, parsedUrl.query.GR, parsedUrl.query.BS, parsedUrl.query.CS, parsedUrl.query.EX);
+        viewerTvProgram(response, results, parsedUrl.query.GR, parsedUrl.query.BS, parsedUrl.query.CS, parsedUrl.query.EX);
     });
 }
