+++ nodejs/requestHandlers/viewTvError.js
@@ -1,8 +1,8 @@
-var viewer = require(__dirname + "/../viewer");
+var viewerViewTvError = require(__dirname + "/../viewer/viewTvError");
 var log = require(__dirname + "/../logger").getLogger();
 
 module.exports = function(response, parsedUrl) {
     log.access.info("Request handler 'viewTvError' was called.");
-    viewer.viewTvError(response);
+    viewerViewTvError(response);
 }
